refactor(customer payment & web route) finish customer payment
